pyproject.toml
[tool.poetry]
name = "rag-system-on-local-machine"
version = "0.1.0"
description = ""
authors = ["Biomega <9dernii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
numpy = "^2.1.0"
transformers = "^4.44.2"
sentence-transformers = "^3.0.1"
python-docx = "^1.1.2"
scipy = "^1.14.1"
python-telegram-bot = "^21.4"
ollama = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
